log: Make sure we use the XSI compliant strerror_r function
We have defined _GNU_SOURCE all over the code to get the GNU specific
asprintf, but this makes a GNU specific strerror_r function avaible,
wich is incompatible with the expected XSI compliant verions.

Defining _GNU_SOURCE only for stdio.h and undefining afterwars is very
hard, because stdio.h is also included by many other headerfiles, so we
would need to make sure, stdio.h included before everything else.

On top of this, there's the problem that windows doesn't support
strerror_r, so we have a workaround which emulates strerror_r
with snprintf.

Though it's easier to use a seperate compilation unit and use a wrapper
function, to ensure the right feature test macros and workarounds
are used.